# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Anotepad
  description: |-
    Danh sách api cho website anotepad.

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://117.1.29.215:4000
paths:
...
  /api/anotepad/note/{noteid}:
    get: 
     tags:
        - note
     summary: ""
     description: Trả về chi tiết note
     operationId: getdetailnote
     parameters:
        - name: noteid
          in: path
          description: id note need to get
          required: true
          schema:
            type: integer    
     responses:
        default:
          description: successful operation
    put:
      tags:
        - note
      summary: ""
      description: update note
      operationId: updatenote 
      parameters:
        - name: noteid
          in: path
          description: id note need to get
          required: true
          schema:
            type: integer  
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
                    type: object
                    properties:
                     title:
                      type: string
                      example: First Note
                     note_type: 
                      type: string
                      example: SimpleNote
                     folder_id:
                      type: integer
                      example: 1
                     status: 
                      type: string
                      example: public
                     content: 
                      type: string
                      example: Dev web anotepad
                     password_access: 
                      type: string
                      example: 20112003
                     password_edit: 
                      type: string
                      example: 20112003
      responses:
        default:
          description: successful operation
    delete:
      tags:
        - note
      summary: Delete note
      description: người dùng xóa note
      operationId: deletenote
      parameters:
        - name: noteid
          in: path
          description: xóa note có id 
          required: true
          schema:
            type: integer
      responses:
        '404':
          description: Note not found
  /api/anotepad/note/accesspermission:
    post:
      tags:
        - note
      summary: ""
      description: Cho phép một người dùng có quyền truy cập vào một note ở chế độ protected
      operationId: accessnote   
      requestBody:
        description: note access object
        content:
          application/json:
            schema:
                    type: object
                    properties:
                     note_id:
                      type: integer
                      example: 1
                     password_access:
                      type: string
                      example: 20112003
      responses:
        default:
          description: successful operation
  /api/anotepad/note/editpermission:
    post:
      tags:
        - note
      summary: ""
      description: Cho phép một người dùng có quyền chỉnh sửa một note ở chế độ protected
      operationId: editnote   
      requestBody:
        description: note edit object
        content:
          application/json:
            schema:
                    type: object
                    properties:
                     note_id:
                      type: integer
                      example: 1
                     password_edit:
                      type: string
                      example: 20112003
      responses:
        default:
          description: successful operation
...